@@ -60,13 +60,13 @@
     "del": "3.0.0",
     "dotenv": "^4.0.0",
     "dotenv-assert": "^3.0.0",
-    "eslint": "^3.19.0",
+    "eslint": "4.3.0",
     "eslint-config-airbnb-base": "^11.2.0",
     "eslint-plugin-import": "^2.3.0",
     "expect": "^1.20.0",
     "gift": "^0.10.0",
     "gulp": "^3.9.0",
-    "gulp-bg": "^0.0.8",
+    "gulp-bg": "^0.0.9",
     "gulp-clean-css": "^3.4.1",
     "gulp-git": "^2.0.0",
     "gulp-mocha": "^4.0.0",
