@@ -16,7 +16,7 @@ module.exports = {
     if (err.name && err.stack) {
       error = `${gutil.colors.red(`${err.plugin}: ${err.name}: `) +
             gutil.colors.bold.red(`${err.message}`)
-            }\n${err.stack}`;
+      }\n${err.stack}`;
     } else {
       error = gutil.colors.red(err);
     }
