@@ -12,18 +12,22 @@ gulp.task('minify', [
   'minify-js',
 ]);
 
-gulp.task('minify-css', () => gulp
+gulp.task('minify-css', () => {
+  gulp
     .src(`${config.dist.cssDir}/${config.dist.cssName}`)
     .pipe(sourcemaps.init())
     .pipe(minifyCss())
     .pipe(rename('styles.min.css'))
     .pipe(sourcemaps.write('./'))
-    .pipe(gulp.dest(config.dist.cssDir)));
+    .pipe(gulp.dest(config.dist.cssDir));
+});
 
-gulp.task('minify-js', () => gulp
+gulp.task('minify-js', () => {
+  gulp
     .src(`${config.dist.jsDir}/${config.dist.jsName}`)
     .pipe(sourcemaps.init())
     .pipe(uglify())
     .pipe(rename('scripts.min.js'))
     .pipe(sourcemaps.write('./'))
-    .pipe(gulp.dest(config.dist.jsDir)));
+    .pipe(gulp.dest(config.dist.jsDir));
+});
