@@ -4,5 +4,5 @@ const config = require('../config');
 
 // runs tests
 gulp.task('test', () => gulp
-    .src(config.tests, { read: false })
-    .pipe(mocha({ reporter: 'spec' })));
+  .src(config.tests, { read: false })
+  .pipe(mocha({ reporter: 'spec' })));
