@@ -1,10 +1,12 @@
-// add fastclick to remove click delay on touch devices
 import attachFastClick from 'fastclick';
+
+import keyboard from './keyboard';
+import homepage from './homepage';
+import application from './application';
+
+homepage();
+keyboard();
+application();
+
+// add fastclick to remove click delay on touch devices
 attachFastClick(document.body);
-
-// load keyboard bindings
-require('./keyboard')();
-
-// load homepage and main app scripts
-require('./homepage')();
-require('./application')();
